requirements.txt
Flask==2.3.3
gunicorn==21.2.0
requests==2.31.0
python-dotenv==1.0.0
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.3
Werkzeug==2.3.7
Pillow==10.4.0
bcrypt==4.0.1
psycopg2-binary
cloudinary==1.36.0
joblib>=1.2.0
pandas>=1.5.0
numpy>=1.24.0
sentence-transformers>=2.2.2
torch>=1.13.0
scikit-learn>=1.2.0
lightgbm>=4.0.0

# Poster Generation
qrcode>=7.4.2
reportlab>=4.0.7

# Free Alert Channels
python-telegram-bot>=20.7

# Optional: Face Recognition (requires dlib + cmake)
# Uncomment below if your platform supports it:
# face-recognition>=1.3.0
# dlib>=19.24.0
